Narrow mdlFile mesh size parameters

+++ src/Graphics/Model/mdlFile.c
@@ -353,7 +353,7 @@ u64 FUN_00323c20(u16 param_1,u16 param_2,u16 param_3,u64 param_4);
 int FUN_00323e10(u32 param_1,u32 param_2,u32 param_3,u32 param_4,int param_5,
 
                  u32 param_6);
-int FUN_00323fb0(u64 param_1,u32 param_2,int param_3,u16 param_4,u64 param_5);
+int FUN_00323fb0(u32 param_1,u32 param_2,int param_3,u16 param_4,u32 param_5);
 u64 FUN_00324160(int param_1);
 void FUN_00324310(int param_1);
  #pragma alias FUN_00324310_i FUN_00324310
@@ -7686,7 +7686,7 @@ int FUN_00323e10(u32 param_1,u32 param_2,u32 param_3,u32 param_4,int param_5,
 // FUN_00323FB0 NONMATCHING
 
 
-int FUN_00323fb0(u64 param_1,u32 param_2,int param_3,u16 param_4,u64 param_5)
+int FUN_00323fb0(u32 param_1,u32 param_2,int param_3,u16 param_4,u32 param_5)
